Get user 's location and display with name
In C++, getline() is a standard library function used to read a full line of input from the user or a file, including spaces, until a newline character is encountered. It’s commonly used when you need to capture entire lines of text, as opposed to cin, which stops reading at whitespace.