Cylinder Volume and Surface Area
This code calculates and displays the volume and surface area of a cylinder. It takes the radius and height of the cylinder as input from the user and uses the appropriate formulas to compute the desired values.

#include <cmath> :  This line includes the cmath header file, which provides mathematical functions like pow for calculating powers.

M_PI :  This is a constant defined in the cmath library that represents the value of pi (π).

pow (radius, 2) :  This expression calculates the square of the radius using the pow function from the cmath library.