Increment and Decrement Operators
This code demonstrates the behavior of post-increment, post-decrement, pre-increment, and pre-decrement operators in C++. It shows how these operators affect the value of a variable before and after the operation.

01) Post-Decrement:

x--  : The value of x is used first, then decremented. So, x is printed as 10, and then it becomes 9.
x--  : Again, the current value of x (which is 9) is used, and then it's decremented to 8.


02) Pre-Decrement:

--y  : The value of y is decremented first, and then the new value (19) is used.


03) Mixed Operations:

z = ++y / ++x   : y is incremented to 20, x is incremented to 9, and then the division is performed.
a = (y++) - (--x)   : y is used first (20), then incremented to 21. x is decremented to 8, and then used. The subtraction is performed.
b = --y + --x   : Both y and x are decremented first, and then their new values (20 and 9) are added.


Remember, post-increment/decrement operators use the current value first, then modify it. Pre-increment/decrement operators modify the value first, then use the new value.