Conditional Statements with if-else in C++
This C++ code demonstrates the use of if-else statements to make decisions based on a user-input number. If the number is greater than 5, the code prints "This number is higher than 5." Otherwise, it prints "This number is lower than 5." Regardless of the condition, the code always prints "The END" at the end.